pyproject.toml
[tool.black]
line-length = 99
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | build
  | dist
)/
'''

[tool.isort]
py_version = 312
line_length = 99
default_section = "THIRDPARTY"
combine_as_imports = true
order_by_type = true
profile = "black"
known_first_party = "lemmy_moderation_dashboard"

[tool.poetry]
name = "lemmy_moderation_dashboard"
version = "0.1.0"
description = "A web-based administration tool to help moderators and admins of Lemmy instances"
authors = ["Raphael Lullis <raphael@mushroomlabs.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pythorhead = "^0.15.6"
requests = "^2"
langdetect = "^1.0"
django = "^5"
django-environ = "^0.10"
django-allauth = "^0.57"
uvicorn = {extras = ["standard"], version = "^0.23"}
django-celery-beat = "^2.5"
psycopg2-binary = "^2.9"
requests-cache = "^1.1"
factory-boy = "^3.3"
django-extensions = "^3.2"
django-model-utils = "^4.3"
bcrypt = "^4.0"
pycryptodome = "^3"

[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4"
pytest-env = "^1.0"
pytest-django = "^4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
